refactor: moved nav.button to forms
