+++ components/styled/form.jsx
@@ -182,3 +182,29 @@ export const option = styled('option', {
 		border: 'true'
 	}
 })
+
+export const button = styled('button', {
+	variants: {
+		theme: {
+			light: {
+				color: 'rgb(0,0,0)',
+				'&:hover': {
+					color: 'rgb(0,0,225)'
+				}
+			},
+			dark: {
+				color: 'rgb(255,255,255)',
+				'&:hover': {
+					color: 'rgb(100,100,220)'
+				}
+			}
+		}
+	},
+	defaultVariants: {
+		color: 'light'
+	},
+	fontWeight: 'bold',
+	textDecoration: 'none',
+	border: 'none',
+	background: 'none'
+})
